% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-class.R
\docType{class}
\name{Action}
\alias{Action}
\title{Classe \code{Action}}
\description{
Cette classe represente le portefeuille des actions de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{ptf}}{est un objet de type \code{\link{data.frame}} contenant les donnees relatives au portefeuille.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
