% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actif-aggregagtion.R
\docType{methods}
\name{aggregation_actif}
\alias{aggregation_actif}
\title{Fonction \code{aggregation_actif}.}
\usage{
aggregation_actif(actif)
}
\arguments{
\item{actif}{est un objet de type \code{\link{Actif}}.}
}
\description{
Cette fonction permet de faire appel aux différentes fonctions permettant d'aggreger les portfeuilles financiers
d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
