% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-aggregagtion.R
\docType{methods}
\name{aggregation_action}
\alias{aggregation_action}
\title{Fonction \code{aggregation_action}.}
\usage{
aggregation_action(action)
}
\arguments{
\item{action}{est un objet de type \code{\link{Action}}.}
}
\description{
Cette fonction permet d'aggreger les model-point pour un portfeuille d'actions.
}
\author{
Damien Tichit pour Sia Partners
}
