% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-aggregation.R
\docType{methods}
\name{aggregation_alm}
\alias{aggregation_alm}
\title{Fonction \code{aggregation_alm}.}
\usage{
aggregation_alm(alm)
}
\arguments{
\item{alm}{est un objet de type \code{\link{ALM}}.}
}
\description{
Cette fonction permet de faire appel aux différentes fonctions permettant d'aggreger les portefeuilles.
}
\author{
Damien Tichit pour Sia Partners
}
