% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-aggregagtion.R
\docType{methods}
\name{aggregation_epargne}
\alias{aggregation_epargne}
\title{Fonction \code{aggregation_epargne}.}
\usage{
aggregation_epargne(epargne)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}
}
\description{
Cette fonction permet d'aggreger les model-point pour un portfeuille d'actions.
}
\author{
Damien Tichit pour Sia Partners
}
