% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-aggregation_1.R
\docType{methods}
\name{aggregation_epargne_1}
\alias{aggregation_epargne_1}
\title{Fonction \code{aggregation_epargne_1}.}
\usage{
aggregation_epargne_1(epargne)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}
}
\description{
Cette fonction permet d'effectuer une premiere aggregation des model-point pour un portefeuille de contrats d'epargne.
Cette aggregation se fait par tmg, sexe, age et anciennete. Elle s'effectue avant le calcul des probas.
}
\author{
Damien Tichit pour Sia Partners
}
