% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-aggregation_2.R
\docType{methods}
\name{aggregation_epargne_2}
\alias{aggregation_epargne_2}
\title{Fonction \code{aggregation_epargne_2}.}
\usage{
aggregation_epargne_2(epargne)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}
}
\description{
Cette fonction permet d'effectuer la derniere aggregation des model-point pour un portefeuille de contrats d'epargne.
Cette aggregation se fait uniquement par tmg. Elle s'effectue apres le calcul des probas.
}
\author{
Damien Tichit pour Sia Partners
}
