% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-aggregagtion.R
\docType{methods}
\name{aggregation_immobilier}
\alias{aggregation_immobilier}
\title{Fonction \code{aggregation_immobilier}.}
\usage{
aggregation_immobilier(immobilier)
}
\arguments{
\item{immobilier}{est un objet de type \code{\link{Immobilier}}.}
}
\description{
Cette fonction permet d'aggreger les model-point pour un portfeuille d'actions.
}
\author{
Damien Tichit pour Sia Partners
}
