% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-aggregation.R
\docType{methods}
\name{aggregation_obligation}
\alias{aggregation_obligation}
\title{Fonction \code{aggregation_obligation}.}
\usage{
aggregation_obligation(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet d'aggreger les model-point pour un portfeuille d'obligations.
}
\author{
Damien Tichit pour Sia Partners
}
