% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-aggregagtion.R
\docType{methods}
\name{aggregation_passif}
\alias{aggregation_passif}
\title{Fonction \code{aggregation_passif}.}
\usage{
aggregation_passif(passif)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}
}
\description{
Cette fonction permet de faire appel aux différentes fonctions permettant d'aggreger les portfeuilles passifs d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
