% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-aggregation_1.R
\docType{methods}
\name{aggregation_passif_1}
\alias{aggregation_passif_1}
\title{Fonction \code{aggregation_passif_1}.}
\usage{
aggregation_passif_1(passif)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}
}
\description{
Cette fonction permet de faire appel aux différentes fonctions permettant d'aggreger les portfeuilles passifs d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
