% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-aggregation_2.R
\docType{methods}
\name{aggregation_passif_2}
\alias{aggregation_passif_2}
\title{Fonction \code{aggregation_passif_2}.}
\usage{
aggregation_passif_2(passif)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}
}
\description{
Cette fonction permet d'effectuer une premiere aggregation des portfeuilles passifs d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
