% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-allocation.R
\docType{methods}
\name{allocation_ptf_actif}
\alias{allocation_ptf_actif}
\title{Fonction \code{allocation_ptf_actif}}
\usage{
allocation_ptf_actif(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}
}
\description{
Cette fonction permet de determiner l'allocation des differents produits financiers.
}
\author{
Damien Tichit pour Sia Partners
}
