% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-allocation_vnc.R
\docType{methods}
\name{allocation_vnc_ptf_actif}
\alias{allocation_vnc_ptf_actif}
\title{Fonction \code{allocation_vnc_ptf_actif}}
\usage{
allocation_vnc_ptf_actif(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{PTFActif}.}
}
\description{
Cette fonction permet de determiner la VNC des differents produits financiers.
}
\author{
Damien Tichit pour Sia Partners
}
