% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-class.R
\docType{class}
\name{ALM}
\alias{ALM}
\title{Classe \code{ALM}}
\description{
Cette classe aggrege un objet \code{\link{System}} ainsi que l'ensemble des hypotheses du modele ALM.
}
\section{Slots}{

\describe{
\item{\code{system}}{est un objet de la classe \code{\link{System}} qui aggregent les actifs et les passifs ainsi que leurs hypothèses.}

\item{\code{hyp_alm}}{est un objet de la classe \code{\link{HypALM}} qui contient les differentes hypotheses du modele ALM.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
