% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-besoin_revalo.R
\docType{methods}
\name{besoin_revalo_epargne}
\alias{besoin_revalo_epargne}
\title{Fonction \code{besoin_revalo_epargne}}
\usage{
besoin_revalo_epargne(epargne, cible)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}

\item{cible}{est un \code{numeric} representant un taux cible.}
}
\description{
Cette fonction permet de calculer les besoins en revalorisation :
\describe{
\item{besoins contractuels}{ : besoins pour assouvir les engagements contractuels, a savoir le tmg.}
\item{besoins cibles}{ : besoins pour atteindre une revalorisation egale au taux cible afin de minimiser les rachats conjoncturels.}
}
}
\author{
Damien Tichit pour Sia Partners
}
