% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFPassif-besoin_revalo.R
\docType{methods}
\name{besoin_revalo_ptf_passif}
\alias{besoin_revalo_ptf_passif}
\title{Fonction \code{besoin_revalo_ptf_passif}}
\usage{
besoin_revalo_ptf_passif(ptf_passif, cible)
}
\arguments{
\item{ptf_passif}{est un objet de type \code{\link{PTFPassif}}.}

\item{cible}{est une \code{list} contenant des elements relatifs a la politique de revalorisation pour les differents passifs.}
}
\description{
Cette fonction permet de calculer les besoins en revalorisation pour les differents PTF passifs d'une compagnie d'assurance :
\describe{
\item{besoins contractuels}{ : besoins pour assouvir les engagements contractuels.}
\item{besoins cibles}{ : besoins pour atteindre une revalorisation cible.}
}
}
\author{
Damien Tichit pour Sia Partners
}
