% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-calc_be.R
\docType{methods}
\name{calc_be}
\alias{calc_be}
\title{Fonction \code{calc_be}.}
\usage{
calc_be(alm, parallel, nb_core)
}
\arguments{
\item{alm}{est un objet de type \code{ALM} contenant l'ensemble des donnees.}

\item{parallel}{est une valeur \code{logical}. Lorsque cet argument est a \code{TRUE}, les calculs sont parallelises.}

\item{nb_core}{est une valeur \code{integer} qui indique le nombre de coeurs utilises lorsque les calculs sont parallelises.}
}
\description{
Cette fonction est une fonction centrale du package. Elle permet en effet de calculer un BEL.
}
\details{
C'est sur cette fonction que s'effectue les boucles sur le nombre de simulations ainsi que sur les annees.
}
\seealso{
Projection sur une annee d'un \code{\link{System}} : \code{\link{proj_1an_system}}.
}
\author{
Damien Tichit pour Sia Partners
}
