% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-calc_be_simu.R
\docType{methods}
\name{calc_be_simu}
\alias{calc_be_simu}
\title{Fonction \code{calc_be_simu}.}
\usage{
calc_be_simu(alm, num_sim)
}
\arguments{
\item{alm}{est un objet de type \code{ALM} contenant l'ensemble des donnees.}

\item{num_sim}{est un \code{integer} representant le numero de simulation sur lequel on travaille.}
}
\description{
Cette fonction est une fonction centrale du package. Elle permet en effet de calculer un BEL pour une simulation donnee.
}
\details{
C'est sur cette fonction que s'effectue les boucles sur le nombre de simulations ainsi que sur les annees.
}
\seealso{
Projection sur une annee d'un \code{\link{System}} : \code{\link{proj_1an_system}}.
}
\author{
Damien Tichit pour Sia Partners
}
