% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-calc_coupon.R,
%   R/Obligation-vieillissement.R
\docType{methods}
\name{calc_coupon}
\alias{calc_coupon}
\alias{calc_coupon}
\title{Fonction \code{calc_coupon}}
\usage{
calc_coupon(obligation)

calc_coupon(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}

\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet de calculer les coupons pour un portfeuille obligataire.

Cette fonction permet de calculer les coupons pour un portfeuille obligataire.
}
\author{
Damien Tichit pour Sia Partners

Damien Tichit pour Sia Partners
}
