% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-calc_dividende.R
\docType{methods}
\name{calc_dividende}
\alias{calc_dividende}
\title{Fonction \code{calc_dividende}}
\usage{
calc_dividende(action)
}
\arguments{
\item{action}{est un objet de type \code{\link{Action}}.}
}
\description{
Cette fonction permet de calculer les dividende pour un portfeuille d'actions.
}
\author{
Damien Tichit pour Sia Partners
}
