% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-calc_loyer.R
\docType{methods}
\name{calc_loyer}
\alias{calc_loyer}
\title{Fonction \code{calc_loyer}}
\usage{
calc_loyer(immobilier)
}
\arguments{
\item{immobilier}{est un objet de type \code{\link{Immobilier}}.}
}
\description{
Cette fonction permet de calculer les loyers pour un portefeuille d'immobiliers.
}
\author{
Damien Tichit pour Sia Partners
}
