% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-calc_proba.R
\docType{methods}
\name{calc_proba_epargne}
\alias{calc_proba_epargne}
\title{Fonction \code{calc_proba_epargne}.}
\usage{
calc_proba_epargne(epargne, hyp_passif, an)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}

\item{hyp_passif}{est un objet de type \code{\link{HypPassif}}.}

\item{an}{est un objet de type \code{integer}.}
}
\description{
Cette fonction permet de completer l'objet contenant les probabilites relatives au portefeuille epargne.
}
\author{
Damien Tichit pour Sia Partners
}
