% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFPassif-calc_proba.R
\docType{methods}
\name{calc_proba_ptf_passif}
\alias{calc_proba_ptf_passif}
\title{Fonction \code{calc_proba_ptf_passif}}
\usage{
calc_proba_ptf_passif(ptf_passif, hyp_passif, an)
}
\arguments{
\item{ptf_passif}{est un objet de type \code{\link{PTFPassif}}.}

\item{hyp_passif}{est un objet de type \code{\link{HypPassif}}.}

\item{an}{est un objet de type \code{integer}.}
}
\description{
Cette fonction permet de calculer et de completer les objets relatifs aux probabilites des differents portefeuilles.
}
\author{
Damien Tichit pour Sia Partners
}
