% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-calc_prod_fin.R
\docType{methods}
\name{calc_prod_fin}
\alias{calc_prod_fin}
\title{Fonction \code{calc_prod_fin}.}
\usage{
calc_prod_fin(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}
}
\description{
Cette fonction permet de calculer les produits financiers pour un portefeuille financier : dividendes, coupons....
}
\author{
Damien Tichit pour Sia Partners
}
