% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabMorta-calc_qx.R
\docType{methods}
\name{calc_qx}
\alias{calc_qx}
\title{Fonction \code{calc_qx}.}
\usage{
calc_qx(tab_morta, age)
}
\arguments{
\item{tab_morta}{est un objet de type \code{\link{TabMorta}}.}

\item{age}{est un \code{integer}.}
}
\description{
Cette fonction permet de calculer des probas de deces pour un age et une table de mortalite donnes.
Il est possible d'indiquer plusieurs ages differents sous forme de vecteur.
}
\author{
Damien Tichit pour Sia Partners
}
