% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RachatConj-calc.R
\docType{methods}
\name{calc_rachat_conj}
\alias{calc_rachat_conj}
\title{Fonction \code{calc_rachat_conj}.}
\usage{
calc_rachat_conj(rachat_conj, tx_cible, tx_serv)
}
\arguments{
\item{rachat_conj}{est un objet de type \code{\link{RachatConj}}.}

\item{tx_cible}{est un \code{numeric}.}

\item{tx_serv}{est un \code{numeric}.}
}
\description{
Cette fonction permet de calculer les taux de rachats conjoncturels.
}
\author{
Damien Tichit pour Sia Partners
}
