% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabRachat-calc_rx.R
\docType{methods}
\name{calc_rx}
\alias{calc_rx}
\title{Fonction \code{calc_rx}.}
\usage{
calc_rx(tab_rachat, age)
}
\arguments{
\item{tab_rachat}{est un objet de type \code{\link{TabRachat}}.}

\item{age}{est un \code{integer}.}
}
\description{
Cette fonction permet de calculer des probas de rachat pour un age et une table de rachat donnes.
Il est possible d'indiquer plusieurs ages differents sous forme de vecteur.
}
\author{
Damien Tichit pour Sia Partners
}
