% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-calc_spread.R
\docType{methods}
\name{calc_spread}
\alias{calc_spread}
\title{Fonction \code{calc_spread}}
\usage{
calc_spread(obligation, yield_curve)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}

\item{yield_curve}{est un vecteur \code{numeric}.}
}
\description{
Cette fonction permet de calculer les spread pour un PTF obligataire.
}
\author{
Damien Tichit pour Sia Partners
}
