% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-calc_tri.R
\docType{methods}
\name{calc_tri}
\alias{calc_tri}
\title{Fonction \code{calc_tri}}
\usage{
calc_tri(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet de calculer les taux actuariels pour un PTF obligataire.
}
\author{
Damien Tichit pour Sia Partners
}
