% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-calcul_pb.R
\docType{methods}
\name{calcul_pb}
\alias{calcul_pb}
\title{Fonction \code{calcul_pb}.}
\usage{
calcul_pb(taux_pb, resultat_fin, resultat_tech)
}
\arguments{
\item{taux_pb}{est une \code{list} contenant les deux taux de pb contractuels.}

\item{resultat_fin}{est une \code{list} contenant les resultats financiers : PMVL, produits financiers.}

\item{resultat_tech}{est une \code{list} contenant le resultat technique : chargements}
}
\description{
Cette fonction permet de determiner la PB a distribuer.
}
\author{
Damien Tichit pour Sia Partners
}
