% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFPassif-calcul_pm.R
\docType{methods}
\name{calcul_pm}
\alias{calcul_pm}
\title{Fonction \code{calcul_pm}}
\usage{
calcul_pm(ptf_passif)
}
\arguments{
\item{ptf_passif}{est un objet de type \code{\link{PTFPassif}}.}
}
\description{
Cette fonction permet de determiner le montant total des PM dans les differents portefeuilles de la compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
