% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-calcul_pt.R
\docType{methods}
\name{calcul_pt}
\alias{calcul_pt}
\title{Fonction \code{calcul_pt}}
\usage{
calcul_pt(passif)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}
}
\description{
Cette fonction permet de determiner le montant total des provisions techniques pour une compagnie d'assurance : PM et PPE
}
\author{
Damien Tichit pour Sia Partners
}
