% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-calcul_quote_part_fp.R
\docType{methods}
\name{calcul_quote_part_fp}
\alias{calcul_quote_part_fp}
\title{Fonction \code{calcul_quote_part_fp}}
\usage{
calcul_quote_part_fp(passif)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}
}
\description{
Cette fonction permet de determiner la proportion que represente les fonds propres sur le passif d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
