% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-calcul_resultat.R
\docType{methods}
\name{calcul_resultat}
\alias{calcul_resultat}
\title{Fonction \code{calcul_resultat}.}
\usage{
calcul_resultat(resultat)
}
\arguments{
\item{resultat}{est une \code{list} contenant les differents elements du resultat.}
}
\description{
Cette fonction permet de calculer le resultat de l'exercice en cours pour une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
