% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-calcul_resultat_fin.R
\docType{methods}
\name{calcul_resultat_fin}
\alias{calcul_resultat_fin}
\title{Fonction \code{calcul_resultat_fin}.}
\usage{
calcul_resultat_fin(resultat_fin)
}
\arguments{
\item{resultat_fin}{est une \code{list} contenant les resultats financiers : PMVR, produits financiers et frais.}
}
\description{
Cette fonction permet de calculer le resultat financier d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
