% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-calcul_resultat_tech.R
\docType{methods}
\name{calcul_resultat_tech}
\alias{calcul_resultat_tech}
\title{Fonction \code{calcul_resultat_tech}.}
\usage{
calcul_resultat_tech(resultat_tech)
}
\arguments{
\item{resultat_tech}{est une \code{list} contenant le resultat technique : chargements}
}
\description{
Cette fonction permet de calculer le resultat technique d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
