% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-calcul_revalo.R
\docType{methods}
\name{calcul_revalo}
\alias{calcul_revalo}
\title{Fonction \code{calcul_revalo}}
\usage{
calcul_revalo(besoin, pb, ppe, revalo_oblig = 0)
}
\arguments{
\item{besoin}{est un \code{numeric}.}

\item{pb}{est un \code{numeric} representant le montant de PB restant.}

\item{ppe}{est un objet de type \code{\link{PPE}}.}

\item{revalo_oblig}{est un \code{numeric}. Il represente un montant de revalorisation obligatoire devant encore etre attribue.}
}
\description{
Cette fonction permet de calculer la revalorisation applique en fonction du besoin, de la PB ainsi que de la PPE.
}
\author{
Damien Tichit pour Sia Partners
}
