% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-dotation.R, R/PPE-dotation_reprise.R
\docType{methods}
\name{dotation_ppe}
\alias{dotation_ppe}
\alias{dotation_ppe}
\title{Fonction \code{dotation_ppe}.}
\usage{
dotation_ppe(ppe, montant)

dotation_reprise_ppe(ppe, montant)
}
\arguments{
\item{ppe}{est un objet de type \code{\link{PPE}}.}

\item{montant}{est un \code{numeric} representant le montant a doter.}

\item{ppe}{est un objet de type \code{\link{PPE}}.}

\item{montant}{est un \code{numeric} representant le montant a doter ou a reprendre.}
}
\description{
Cette fonction permet de doter la PPE. Le montant est dote sur la 1ere annee de reserve.

Cette fonction permet de reprendre ou doter un montant a la PPE.
\describe{
\item{dotation}{ : Appel de la fonction \code{\link{dotation_ppe}} ;}
\item{reprise}{ : Appel de la fonction \code{\link{reprise_ppe}}.}
}
}
\author{
Damien Tichit pour Sia Partners

Damien Tichit pour Sia Partners
}
