% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG-class.R
\docType{class}
\name{ESG}
\alias{ESG}
\title{Classe \code{ESG}}
\description{
Cette classe aggrege l'ensemble des donnees relatives  a un ESG. Les donnees sont renseignees par simulation ainsi que par annee de projection.
}
\section{Slots}{

\describe{
\item{\code{coef_actu}}{est un \code{data.table} : coefficients d'actualisation.}

\item{\code{ctz_nom}}{est un \code{data.table} : taux zero coupon reels.}

\item{\code{ctz_reel}}{est un \code{data.table} : taux zero coupon nominaux.}

\item{\code{eq_dividends}}{est un \code{data.table} : dividendes verses.}

\item{\code{eq_index}}{est un \code{data.table} : log-rendements du PTF action.}

\item{\code{im_index}}{est un \code{data.table} : log-rendements du PTF immobilier}

\item{\code{im_loyer}}{est un \code{data.table} : loyers verses.}

\item{\code{inflation}}{est un \code{data.table} : taux d'inflation.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
