% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-eval_coupon.R
\docType{methods}
\name{eval_coupon}
\alias{eval_coupon}
\title{Fonction \code{eval_coupon}}
\usage{
eval_coupon(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet de calculer les coupons pour un portfeuille obligataire.
}
\author{
Damien Tichit pour Sia Partners
}
