% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-eval_dividende.R
\docType{methods}
\name{eval_dividende}
\alias{eval_dividende}
\title{Fonction \code{eval_dividende}}
\usage{
eval_dividende(action, div)
}
\arguments{
\item{action}{est un objet de type \code{\link{Action}}.}

\item{div}{est un \code{numeric} represente la proportion de l'encours action versee au titre des dividendes.}
}
\description{
Cette fonction permet de calculer les dividendes pour un portfeuille d'actions.
}
\author{
Damien Tichit pour Sia Partners
}
