% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-eval_loyer.R
\docType{methods}
\name{eval_loyer}
\alias{eval_loyer}
\title{Fonction \code{eval_loyer}}
\usage{
eval_loyer(immobilier)
}
\arguments{
\item{immobilier}{est un objet de type \code{\link{Immobilier}}.}
}
\description{
Cette fonction permet de calculer les loyers pour un portefeuille d'immobiliers.
}
\author{
Damien Tichit pour Sia Partners
}
