% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-extract_pmvl.R
\docType{methods}
\name{extract_pmvl_action}
\alias{extract_pmvl_action}
\title{Fonction \code{extract_pmvl_action}}
\usage{
extract_pmvl_action(action)
}
\arguments{
\item{action}{est un objet de type \code{Action}.}
}
\description{
Cette fonction permet de determiner les PMVL pouvant etre realises sur un PTF d'actions.
}
\author{
Damien Tichit pour Sia Partners
}
