% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-extract_pmvl.R
\docType{methods}
\name{extract_pmvl_immobilier}
\alias{extract_pmvl_immobilier}
\title{Fonction \code{extract_pmvl_immobilier}}
\usage{
extract_pmvl_immobilier(immobilier)
}
\arguments{
\item{immobilier}{est un objet de type \code{Immobilier}.}
}
\description{
Cette fonction permet de determiner les PMVL pouvant etre realises sur un PTF immobilier.
}
\author{
Damien Tichit pour Sia Partners
}
