% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-extract_pmvl.R
\docType{methods}
\name{extract_pmvl_obligation}
\alias{extract_pmvl_obligation}
\title{Fonction \code{extract_pmvl_obligation}}
\usage{
extract_pmvl_obligation(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{Obligation}.}
}
\description{
Cette fonction permet de determiner les PMVL pouvant etre realises sur un PTF obligataire.
}
\author{
Damien Tichit pour Sia Partners
}
