% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-extract_pmvl.R
\docType{methods}
\name{extract_pmvl_ptf_actif}
\alias{extract_pmvl_ptf_actif}
\title{Fonction \code{extract_pmvl_ptf_actif}}
\usage{
extract_pmvl_ptf_actif(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{PTFActif}.}
}
\description{
Cette fonction permet de determiner les PMVL pouvant etre realises sur les differents portefeuille.
}
\author{
Damien Tichit pour Sia Partners
}
