% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FondsPropres-class.R
\docType{class}
\name{FondsPropres}
\alias{FondsPropres}
\title{Classe \code{FondsPropres}}
\description{
Cette classe aggrege l'ensemble des donnees relatives au passif de la compagnie d'assurance : hypotheses, portefeuille, provisions
}
\section{Slots}{

\describe{
\item{\code{capitaux_propres}}{est un \code{numeric} representant le montant des capitaux propres}

\item{\code{report_a_nouveau}}{est un \code{numeric} representant le montant du report a nouveau.}

\item{\code{resultat_exercice}}{est un \code{numeric} representant le resultat de l'exercice en cours.}

\item{\code{dette}}{est un \code{numeric}.}

\item{\code{hypotheses}}{est une \code{list}.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
