% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-gestion.R
\docType{methods}
\name{gestion_action}
\alias{gestion_action}
\title{Fonction \code{gestion_action}}
\usage{
gestion_action(action, esg_simu, an)
}
\arguments{
\item{action}{est un objet de type \code{\link{Action}}.}

\item{esg_simu}{est une \code{list} contenant les rendements et les dividendes.}

\item{an}{est un \code{integer} reprensentant l'annee sur laquelle on travaille.}
}
\description{
Cette fonction permet de gerer un portfeuille action : recolte des dividendes, recalcul des VM
}
\author{
Damien Tichit pour Sia Partners
}
