% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-gestion_capitaux_propres.R
\docType{methods}
\name{gestion_capitaux_propres}
\alias{gestion_capitaux_propres}
\title{Fonction \code{gestion_capitaux_propres}}
\usage{
gestion_capitaux_propres(passif, result_fin)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}

\item{result_fin}{est un \code{numeric} representant le resultat financier total.}
}
\description{
Cette fonction permet de gerer les capitaux propres : ajout d'une part du resultat fin, retirer les dividendes.
}
\author{
Damien Tichit pour Sia Partners
}
