% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actif-gestion_fin_projection.R
\docType{methods}
\name{gestion_fin_projection_actif}
\alias{gestion_fin_projection_actif}
\title{Fonction \code{gestion_fin_projection_actif}}
\usage{
gestion_fin_projection_actif(actif)
}
\arguments{
\item{actif}{est un objet de type \code{\link{Actif}}.}
}
\description{
Cette fonction permet de gerer la fin de projection du modele sur l'actif de la compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
