% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-gestion_fin_projection.R
\docType{methods}
\name{gestion_fin_projection_passif}
\alias{gestion_fin_projection_passif}
\title{Fonction \code{gestion_fin_projection_passif}}
\usage{
gestion_fin_projection_passif(passif)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}
}
\description{
Cette fonction permet de gerer la fin de projection du modele sur le passif de la compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
