% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-gestion_fin_projection.R
\docType{methods}
\name{gestion_fin_projection_system}
\alias{gestion_fin_projection_system}
\title{Fonction \code{gestion_fin_projection_system}}
\usage{
gestion_fin_projection_system(system)
}
\arguments{
\item{system}{est un objet de type \code{\link{System}}.}
}
\description{
Cette fonction permet de gerer la fin de projection de la compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
