% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FondsPropres-gestion.R
\docType{methods}
\name{gestion_fonds_propres}
\alias{gestion_fonds_propres}
\title{Fonction \code{gestion_fonds_propres}}
\usage{
gestion_fonds_propres(fp, resultat, emprunt)
}
\arguments{
\item{fp}{est un objet de type \code{FondsPropres}.}

\item{resultat}{est un \code{numeric} representant le resultat de l'annee.}

\item{emprunt}{est un \code{numeric}.}
}
\description{
Cette fonction permet de gerer les fonds propres.
}
\author{
Damien Tichit pour Sia Partners
}
