% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-gestion.R
\docType{methods}
\name{gestion_obligation}
\alias{gestion_obligation}
\title{Fonction \code{gestion_obligation}}
\usage{
gestion_obligation(obligation, ctz_nom, an)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}

\item{ctz_nom}{est un objet de type \code{data.frame} contenant la courbe des taux.}

\item{an}{est un \code{integer} reprensentant l'annee sur laquelle on travaille.}
}
\description{
Cette fonction permet de gerer un portfeuille obligation : recolte des coupons, recalcul des VM, VNC, VR, vieillissement...
}
\author{
Damien Tichit pour Sia Partners
}
