% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-gestion.R
\docType{methods}
\name{gestion_ptf_actif}
\alias{gestion_ptf_actif}
\title{Fonction \code{gestion_ptf_actif}.}
\usage{
gestion_ptf_actif(ptf_actif, hyp_actif, an)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}

\item{hyp_actif}{est un objet de type \code{\link{HypActif}}.}

\item{an}{est un \code{integer} reprensentant l'annee sur laquelle on travaille.}
}
\description{
Cette fonction permet de gerer les differents ptf actif d'une compagnie d'assurance : recolte des produits financiers, recalcul des VM, vieillissement...
}
\author{
Damien Tichit pour Sia Partners
}
