% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypALM-class.R
\docType{class}
\name{HypALM}
\alias{HypALM}
\title{Classe \code{HypALM}}
\description{
Cette classe aggrege l'ensemble des hypotheses relatives au modele ALM.
}
\section{Slots}{

\describe{
\item{\code{nb_simu}}{est un \code{integer} representant le nombre de simulation souhaite pour calculer le BEL.}

\item{\code{an_proj}}{est un \code{integer} representant le nombre d'annees de projection.}

\item{\code{tsr}}{est un \code{data.frame} contenant les taux sans risque avec Volatility Adjustment.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
