% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-class.R
\docType{class}
\name{Immobilier}
\alias{Immobilier}
\title{Classe \code{Immobilier}}
\description{
Cette classe represente le portefeuille des immobilier de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{ptf}}{est un objet de type \code{\link{data.frame}} contenant les donnees relatives au portefeuille.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
