% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-load.R
\docType{methods}
\name{load_action}
\alias{load_action}
\title{Fonction \code{load_action}.}
\usage{
load_action(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Action}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Action}} necessite des donnees presentes dans un fichier nomme \code{Actions.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
