% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-load.R
\docType{methods}
\name{load_alm}
\alias{load_alm}
\title{Fonction \code{load_alm}.}
\usage{
load_alm(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaire
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{ALM}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{ALM}} necessite la creation de deux autres objets : \code{\link{System}} et \code{\link{HypALM}}.
}
\seealso{
Construction d'un objet de type \code{\link{System}} : \code{\link{load_system}}.

Construction d'un objet de type \code{\link{HypALM}} : \code{\link{load_hyp_alm}}.
}
\author{
Damien Tichit pour Sia Partners
}
