% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-load.R
\docType{methods}
\name{load_epargne}
\alias{load_epargne}
\title{Fonction \code{load_epargne}.}
\usage{
load_epargne(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Epargne}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Epargne}} necessite des donnees presentes dans un fichier nomme \code{Epargne.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
