% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG-load.R
\docType{methods}
\name{load_esg}
\alias{load_esg}
\title{Fonction \code{load_esg}.}
\usage{
load_esg(address, avec_va = TRUE)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}

\item{avec_va}{est un \code{logical} indiquant si les donnees doivent etre chargees sur le jeu de donnees avec Volatility Adjustment. Par defaut, la valeur est a \code{TRUE}.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{ESG}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\author{
Damien Tichit pour Sia Partners
}
