% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FondsPropres-load.R
\docType{methods}
\name{load_fonds_propres}
\alias{load_fonds_propres}
\title{Fonction \code{load_fonds_propres}.}
\usage{
load_fonds_propres(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{FondsPropres}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\author{
Damien Tichit pour Sia Partners
}
