% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypActif-load.R
\docType{methods}
\name{load_hyp_actif}
\alias{load_hyp_actif}
\title{Fonction \code{load_hyp_actif}.}
\usage{
load_hyp_actif(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{HypActif}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{HypActif}} necessite la creation d'un autre objet : \code{\link{PTFCible}}.
}
\seealso{
Construction d'un objet de type \code{\link{PTFCible}} : \code{\link{load_ptf_cible}}.
}
\author{
Damien Tichit pour Sia Partners
}
