% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypALM-load.R
\docType{methods}
\name{load_hyp_alm}
\alias{load_hyp_alm}
\title{Fonction \code{load_hyp_alm}.}
\usage{
load_hyp_alm(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{HypALM}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{HypALM}} necessite plusieurs parametres presents dans un fichier nomme \code{Hypotheses_ALM.csv}.
}
\seealso{
Chargement d'un objet \code{\link{ESG}} : \code{\link{load_esg}}
}
\author{
Damien Tichit pour Sia Partners
}
