% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypPassif-load.R
\docType{methods}
\name{load_hyp_passif}
\alias{load_hyp_passif}
\title{Fonction \code{load_hyp_passif}.}
\usage{
load_hyp_passif(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{HypPassif}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{HypPassif}} necessite la creation de deux autres objets differents : \code{\link{TabMorta}} et \code{\link{TabRachat}}.
}
\seealso{
Construction d'un objet de type \code{\link{TabMorta}} : \code{\link{load_tab_morta}}.

Construction d'un objet de type \code{\link{TabRachat}} : \code{\link{load_tab_rachat}}.
}
\author{
Damien Tichit pour Sia Partners
}
