% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-load_cible.R
\docType{methods}
\name{load_immobilier_cible}
\alias{load_immobilier_cible}
\title{Fonction \code{load_immobilier_cible}.}
\usage{
load_immobilier_cible(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Immobilier}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Immobilier}} necessite des donnees presentes dans un fichier nomme \code{Immobilier.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
