% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-load.R
\docType{methods}
\name{load_obligation}
\alias{load_obligation}
\title{Fonction \code{load_obligation}.}
\usage{
load_obligation(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Obligation}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Obligation}} necessite des donnees presentes dans un fichier nomme \code{Obligations.csv}.

Cette fonction permet egalement d'aggreger les donnees : 1 obligation par maturite residuelle.
}
\author{
Damien Tichit pour Sia Partners
}
