% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-load.R
\docType{methods}
\name{load_ppe}
\alias{load_ppe}
\title{Fonction \code{load_ppe}.}
\usage{
load_ppe(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{PPE}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{PPE}} necessite des donnees presentes dans les fichiers nommes \code{PPE.csv} et \code{TM-H.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
