% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE-load.R
\docType{methods}
\name{load_pre}
\alias{load_pre}
\title{Fonction \code{load_pre}.}
\usage{
load_pre(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{PRE}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{PRE}} necessite des donnees presentes dans les fichiers nommes \code{pre.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
