% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Provision-load.R
\docType{methods}
\name{load_provision}
\alias{load_provision}
\title{Fonction \code{load_provision}.}
\usage{
load_provision(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Provision}}.
Les donnees auront ete prealablement disposees dans une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Provision}} necessite la creation de deux autres objets : \code{\link{PPE}} et \code{\link{ReserveCapi}}.
}
\seealso{
Construction d'un objet de type \code{\link{PPE}} : \code{\link{load_ppe}}.

Construction d'un objet de type \code{\link{ReserveCapi}} : \code{\link{load_reserve_capi}}.
}
\author{
Damien Tichit pour Sia Partners
}
