% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-load.R
\docType{methods}
\name{load_ptf_actif}
\alias{load_ptf_actif}
\title{Fonction \code{load_ptf_actif}.}
\usage{
load_ptf_actif(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Actif}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Actif}} necessite la creation de trois autres objets : \code{\link{Action}}, \code{\link{Obligation}} et \code{\link{Tresorerie}}.
}
\seealso{
Construction d'un objet de type \code{\link{Action}} : \code{\link{load_action}}.

Construction d'un objet de type \code{\link{Obligation}} : \code{\link{load_obligation}}.

Construction d'un objet de type \code{\link{Tresorerie}} : \code{\link{load_tresorerie}}.
}
\author{
Damien Tichit pour Sia Partners
}
