% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFPassif-load.R
\docType{methods}
\name{load_ptf_passif}
\alias{load_ptf_passif}
\title{Fonction \code{load_ptf_passif}.}
\usage{
load_ptf_passif(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{PTFPassif}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{PTFPassif}} necessite la creation d'un autre objet : \code{\link{Epargne}}.
}
\seealso{
Construction d'un objet de type \code{\link{Epargne}} : \code{\link{load_epargne}}.
}
\author{
Damien Tichit pour Sia Partners
}
