% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RachatConj-load.R
\docType{methods}
\name{load_rachat_conj}
\alias{load_rachat_conj}
\title{Fonction \code{load_rachat_conj}.}
\usage{
load_rachat_conj(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{RachatConj}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{RachatConj}} necessite des donnees presentes dans un fichier nommes \code{Rachats_conjoncturels.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
