% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReserveCapi-load.R
\docType{methods}
\name{load_reserve_capi}
\alias{load_reserve_capi}
\title{Fonction \code{load_reserve_capi}.}
\usage{
load_reserve_capi(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{ReserveCapi}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{ReserveCapi}} necessite des donnees presentes dans les fichiers nommes \code{reserve_capi.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
