% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-load.R
\docType{methods}
\name{load_system}
\alias{load_system}
\title{Fonction \code{load_system}.}
\usage{
load_system(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{System}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{System}} necessite la creation de deux autres objets : \code{\link{Actif}} et \code{\link{Passif}}.
}
\seealso{
Construction d'un objet de type \code{\link{Actif}} : \code{\link{load_actif}}.

Construction d'un objet de type \code{\link{Passif}} : \code{\link{load_passif}}.
}
\author{
Damien Tichit pour Sia Partners
}
