% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabMorta-load.R
\docType{methods}
\name{load_tab_morta}
\alias{load_tab_morta}
\title{Fonction \code{load_tab_morta}.}
\usage{
load_tab_morta(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{TabMorta}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{TabMorta}} necessite des donnees presentes dans les fichiers nommes \code{TM-F.csv} et \code{TM-H.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
