% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabRachat-load.R
\docType{methods}
\name{load_tab_rachat}
\alias{load_tab_rachat}
\title{Fonction \code{load_tab_rachat}.}
\usage{
load_tab_rachat(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{TabRachat}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{TabRachat}} necessite des donnees presentes dans un fichier nomme \code{Rachat.csv}.
}
\author{
Damien Tichit pour Sia Partners
}
