% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\docType{methods}
\name{pna.omit}
\alias{pna.omit}
\title{Fonction \code{pna.omit}}
\usage{
pna.omit(x, y)
}
\arguments{
\item{x}{premier vecteur.}

\item{y}{second vecteur.}
}
\description{
Cette fonction, prenant deux vecteurs, sort un seul vecteur sans NAs.
}
\author{
Damien Tichit pour Sia Partners
}
