% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-class.R
\docType{class}
\name{PPE}
\alias{PPE}
\title{Classe \code{PPE}}
\description{
Cette classe represente le Provision pour Participation aux Excedents.
}
\section{Slots}{

\describe{
\item{\code{ppe}}{est un \code{\link{numeric}} contenant les montants dotes sur les huits dernieres annees.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
