% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE-class.R
\docType{class}
\name{PRE}
\alias{PRE}
\title{Classe \code{PRE}}
\description{
Cette classe permet de modeliser la PRE.
}
\section{Slots}{

\describe{
\item{\code{montant}}{est un \code{numeric}.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
