% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbaEpargne-class.R
\docType{class}
\name{ProbaEpargne}
\alias{ProbaEpargne}
\title{Classe \code{ProbaEpargne}}
\description{
Cette classe permet de stocker les differentes probas relatives a un portfeuille epargne : deces et rachats.
}
\section{Slots}{

\describe{
\item{\code{deces_contr}}{est un objet de type \code{data.frame}.}

\item{\code{deces_pm}}{est un objet de type \code{data.frame}.}

\item{\code{rachat_part}}{est un objet de type \code{data.frame}.}

\item{\code{rachat_tot_contr}}{est un objet de type \code{data.frame}.}

\item{\code{rachat_tot_pm}}{est un objet de type \code{data.frame}.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
