% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-proj_1an.R
\docType{methods}
\name{proj_1an_action}
\alias{proj_1an_action}
\title{Fonction \code{proj_1an_action}}
\usage{
proj_1an_action(action)
}
\arguments{
\item{action}{est un objet de type \code{\link{Action}}.}
}
\description{
Cette fonction permet de projeter horizon 1 an un portefeuille d'actions. Elle calcule notamment les dividendes.
}
\author{
Damien Tichit pour Sia Partners
}
