% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-proj_1an.R
\docType{methods}
\name{proj_1an_epargne}
\alias{proj_1an_epargne}
\title{Fonction \code{proj_1an_epargne}.}
\usage{
proj_1an_epargne(epargne, hyp_passif)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}

\item{hyp_passif}{est un objet de type \code{\link{HypPassif}}.}
}
\description{
Cette fonction permet de projeter horizon 1 an un portefeuille de contrats epragnes.
}
\author{
Damien Tichit pour Sia Partners
}
