% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-proj_1an.R
\docType{methods}
\name{proj_1an_immobilier}
\alias{proj_1an_immobilier}
\title{Fonction \code{proj_1an_immobilier}}
\usage{
proj_1an_immobilier(immobilier)
}
\arguments{
\item{immobilier}{est un objet de type \code{\link{Immobilier}}.}
}
\description{
Cette fonction permet de projeter horizon 1 an un portefeuille d'immobiliers. Elle calcule notamment les loyers.
}
\author{
Damien Tichit pour Sia Partners
}
