% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-proj_1an.R
\docType{methods}
\name{proj_1an_obligation}
\alias{proj_1an_obligation}
\title{Fonction \code{proj_1an_obligation}}
\usage{
proj_1an_obligation(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet de projeter horizon 1 an un portefeuille d'obligations. Elle permet notamme de caluler les coupons, de vieillir le
portfeuille ainsi que de vendre les obligations arrivees a maturite.
}
\author{
Damien Tichit pour Sia Partners
}
