% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-proj_1an.R
\docType{methods}
\name{proj_1an_ptf_actif}
\alias{proj_1an_ptf_actif}
\title{Fonction \code{proj_1an_ptf_actif}.}
\usage{
proj_1an_ptf_actif(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}
}
\description{
Cette fonction permet de projeter horizon 1 an un portfeuille financier : gestion des diffrents actifs.
}
\author{
Damien Tichit pour Sia Partners
}
