% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-proj_1an.R
\docType{methods}
\name{proj_1an_system}
\alias{proj_1an_system}
\title{Fonction \code{proj_1an_system}.}
\usage{
proj_1an_system(system, an)
}
\arguments{
\item{system}{est un objet de type \code{System}.}

\item{an}{est un \code{integer}.}
}
\description{
Cette fonction permet de projeter a horizon 1 an un objet \code{\link{System}}.
}
\seealso{
Projection des passifs : \code{\link{proj_1an_passif}}

Projection des actifs : \code{\link{proj_1an_actif}}
}
\author{
Damien Tichit pour Sia Partners
}
