% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Provision-class.R
\docType{class}
\name{Provision}
\alias{Provision}
\title{Classe \code{Provision}}
\description{
Cette classe aggrege les differentes provisions relatives au passif d'une compagnie d'assurance : PPE, Reserve de Capitalisation
}
\section{Slots}{

\describe{
\item{\code{ppe}}{est un objet de la classe \code{\link{PPE}}.}

\item{\code{reserve_capi}}{est un objet de la classe \code{\link{ReserveCapi}}.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
