% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\docType{methods}
\name{psum}
\alias{psum}
\title{Fonction \code{psum}}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{...}{les differnts vecteurs.}

\item{na.rm}{est un \code{logical} qui, lorsqu'il est a TRUE, permet a la somme de ne pas prendre en compte les NAs.}
}
\description{
Cette fonction permet de faire la somme, sur plusieurs vecteurs, element par element, et pouvant supprimer les NAs.
}
\author{
Damien Tichit pour Sia Partners
}
