% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-class.R
\docType{class}
\name{PTFActif}
\alias{PTFActif}
\title{Classe \code{PTFActif}}
\description{
Cette classe represente le portfeuille financier de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{action}}{est un objet de la classe \code{\link{Action}} representant le portfeuille action.}

\item{\code{obligation}}{est un objet de la classe \code{\link{Obligation}} representant le portfeuille obligation.}

\item{\code{tresorerie}}{est un objet de la classe \code{\link{Tresorerie}} representant le portfeuille tresorerie.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
