% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFCible-class.R
\docType{class}
\name{PTFCible}
\alias{PTFCible}
\title{Classe \code{PTFCible}}
\description{
Cette classe represente le portfeuille financier cible de la compagnie d'assurance dans le cadre de
}
\section{Slots}{

\describe{
\item{\code{obligation}}{est un objet de la classe \code{\link{Obligation}} representant le portfeuille cible obligation.}

\item{\code{alloc_cible}}{est un objet \code{data.frame} indiquant l'allocation cible pour chaque produit.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
