% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFPassif-class.R
\docType{class}
\name{PTFPassif}
\alias{PTFPassif}
\title{Classe \code{PTFPassif}}
\description{
Cette classe represente le portfeuille passif de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{epargne}}{est un objet de la classe \code{\link{Epargne}} representant le portfeuille epargne.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
