% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RachatConj-class.R
\docType{class}
\name{RachatConj}
\alias{RachatConj}
\title{Classe \code{RachatConj}}
\description{
Cette classe permet de modeliser les rachats conjoncturels.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{est un \code{numeric}}

\item{\code{beta}}{est un \code{numeric}}

\item{\code{gamma}}{est un \code{numeric}}

\item{\code{delta}}{est un \code{numeric}}

\item{\code{RCmin}}{est un \code{numeric}}

\item{\code{RCmax}}{est un \code{numeric}}

\item{\code{repartition}}{est une \code{list}}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
