% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-realisation_pvl.R
\docType{methods}
\name{realisation_pvl_action}
\alias{realisation_pvl_action}
\title{Fonction \code{realisation_pvl_action}}
\usage{
realisation_pvl_action(action, montant)
}
\arguments{
\item{action}{est un objet de type \code{Action}.}

\item{montant}{est un \code{numeric} indiquant le montant de plus value devant etre realise.}
}
\description{
Cette fonction permet de realiser des plus values lattentes.
}
\author{
Damien Tichit pour Sia Partners
}
