% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-realisation_pvl.R
\docType{methods}
\name{realisation_pvl_immobilier}
\alias{realisation_pvl_immobilier}
\title{Fonction \code{realisation_pvl_immobilier}}
\usage{
realisation_pvl_immobilier(immobilier, montant)
}
\arguments{
\item{immobilier}{est un objet de type \code{Immobilier}.}

\item{montant}{est un \code{numeric} indiquant le montant de plus value devant etre realise.}
}
\description{
Cette fonction permet de realiser des plus values lattentes.
}
\author{
Damien Tichit pour Sia Partners
}
