% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-realisation_pvl.R
\docType{methods}
\name{realisation_pvl_obligation}
\alias{realisation_pvl_obligation}
\title{Fonction \code{realisation_pvl_obligation}}
\usage{
realisation_pvl_obligation(obligation, montant)
}
\arguments{
\item{obligation}{est un objet de type \code{Obligation}.}

\item{montant}{est un \code{numeric} indiquant le montant de plus value devant etre realise.}
}
\description{
Cette fonction permet de realiser des plus values lattentes pour un portefeuille obligataire.
}
\author{
Damien Tichit pour Sia Partners
}
