% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-realisation_pvl.R
\docType{methods}
\name{realisation_pvl_ptf_actif}
\alias{realisation_pvl_ptf_actif}
\title{Fonction \code{realisation_pvl_ptf_actif}}
\usage{
realisation_pvl_ptf_actif(ptf_actif, montant)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{PTFActif}.}

\item{montant}{est un \code{numeric} indiquant le montant de plus value devant etre realise.}
}
\description{
Cette fonction permet de realiser des plus values lattentes.
}
\author{
Damien Tichit pour Sia Partners
}
