% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-reprise.R
\docType{methods}
\name{reprise_ppe}
\alias{reprise_ppe}
\title{Fonction \code{reprise_ppe}.}
\usage{
reprise_ppe(ppe, montant)
}
\arguments{
\item{ppe}{est un objet de type \code{\link{PPE}}.}

\item{montant}{est un \code{numeric} representant le montant a reprendre.}
}
\description{
Cette fonction permet de reprendre un montant sur la PPE. Le montant est prioritairement repris sur les plus vieilles dotations.
}
\author{
Damien Tichit pour Sia Partners
}
