% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-reprise_ppe_8ans.R
\docType{methods}
\name{reprise_ppe_8ans}
\alias{reprise_ppe_8ans}
\title{Fonction \code{reprise_ppe_8ans}.}
\usage{
reprise_ppe_8ans(ppe)
}
\arguments{
\item{ppe}{est un objet de type \code{\link{PPE}}.}
}
\description{
Cette fonction permet de reprendre le montant dotee 8 annees auparavant. Elle met egalement a 0 l'element du vecteur correspondant a la PPE dotee 8 ans auparavant.
}
\author{
Damien Tichit pour Sia Partners
}
