% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReserveCapi-class.R
\docType{class}
\name{ReserveCapi}
\alias{ReserveCapi}
\title{Classe \code{ReserveCapi}}
\description{
Cette classe represente la Reserve de Capitalisation.
}
\section{Slots}{

\describe{
\item{\code{montant}}{est un \code{numeric} representant le capital present dans la reserve.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
