% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action-revalo.R
\docType{methods}
\name{revalo_action}
\alias{revalo_action}
\title{Fonction \code{revalo_action}}
\usage{
revalo_action(action, log_rdt)
}
\arguments{
\item{action}{est un objet de type \code{\link{Action}}.}

\item{log_rdt}{est un \code{numeric} representant le log-rendement du PTF.}
}
\description{
Cette fonction permet de revaloriser le portefeuille action. Calcule egalement les plus ou moins values latentes.
}
\author{
Damien Tichit pour Sia Partners
}
