% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-revalo.R
\docType{methods}
\name{revalo_epargne}
\alias{revalo_epargne}
\title{Fonction \code{revalo_epargne}}
\usage{
revalo_epargne(epargne, revalo_cible, revalo_supp, cible)
}
\arguments{
\item{epargne}{est un objet de type \code{\link{Epargne}}.}

\item{revalo_cible}{est un \code{numeric} representant le montant de revalorisation cible a distibuer.}

\item{revalo_supp}{est un \code{numeric} representant le montant de revalorisation supplementaire a distibuer.}

\item{cible}{est un \code{numeric} representant un taux cible.}
}
\description{
Cette fonction permet de revaloriser un portefeuille Epargne. La distribution se fait de la faca suivante :
\describe{
\item{Revalorisation cible}{ : la revalorisation s'effectue de telle façon a ce que tous les MP aient la meme revalorisation et en se rapprochant du taux cible.}
\item{Revalorisation supplementaire}{ : la revalorisation s'effectue au proportionnellement aux PM.}
}
Les besoins reglementaires sont tout d'abord decompté du montant devant etre obligatoirement distribue.
}
\author{
Damien Tichit pour Sia Partners
}
