% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-revalo.R
\docType{methods}
\name{revalo_obligation}
\alias{revalo_obligation}
\title{Fonction \code{revalo_obligation}}
\usage{
revalo_obligation(obligation, yield_curve)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}

\item{yield_curve}{est un \code{numeric} contenant les prix zero-coupon.}
}
\description{
Cette fonction permet de revaloriser les differentes obligations d'un portefeuille obligataire.
Elle calcule egalement les plus ou moins value latentes (PMVL) engendrees.
}
\author{
Damien Tichit pour Sia Partners
}
