% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-revalo_cible.R
\docType{methods}
\name{revalo_obligation_cible}
\alias{revalo_obligation_cible}
\title{Fonction \code{revalo_obligation_cible}}
\usage{
revalo_obligation_cible(obligation, yield_curve)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}

\item{yield_curve}{est un \code{numeric} contenant les prix zero-coupon.}
}
\description{
Cette fonction permet de revaloriser les differentes obligations d'un portefeuille obligataire cible :
calcul des nouveaux coupons et mise a jour de la valeur de marche.
}
\author{
Damien Tichit pour Sia Partners
}
