% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-revalo.R
\docType{methods}
\name{revalo_passif}
\alias{revalo_passif}
\title{Fonction \code{revalo_passif}}
\usage{
revalo_passif(passif, pb)
}
\arguments{
\item{passif}{est un objet de type \code{\link{Passif}}.}

\item{pb}{est un \code{numeric} representant le montant de PB a distribuer.}
}
\description{
Cette fonction permet de revaloriser le passif d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
