% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFCible-revalo.R
\docType{methods}
\name{revalo_ptf_cible}
\alias{revalo_ptf_cible}
\title{Fonction \code{revalo_ptf_cible}}
\usage{
revalo_ptf_cible(ptf_cible, esg, an)
}
\arguments{
\item{ptf_cible}{est un objet de type \code{\link{PTFCible}}.}

\item{esg}{est une \code{list}.}

\item{an}{est un \code{integer} reprensentant l'annee sur laquelle on travaille.}
}
\description{
Cette fonction permet de revaloriser les portfeuilles cibles : mise a jour des valeurs de marche et des
}
\author{
Damien Tichit pour Sia Partners
}
