% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFPassif-revalo.R
\docType{methods}
\name{revalo_ptf_passif}
\alias{revalo_ptf_passif}
\title{Fonction \code{revalo_ptf_passif}}
\usage{
revalo_ptf_passif(ptf_passif, revalo_cible, revalo_supp, cible)
}
\arguments{
\item{ptf_passif}{est un objet de type \code{\link{PTFPassif}}.}

\item{revalo_cible}{est une \code{list} contenant les montants de revalorisation cibles a distibuer par produit.}

\item{revalo_supp}{est une \code{list} contenant les montants de revalorisation supplementaire a distibuer par produit.}

\item{cible}{est une \code{list} contenant des elements relatifs a la politique de revalorisation pour les differents passifs.}
}
\description{
Cette fonction permet de revaloriser les differents PTF passifs d'une compagnie d'assurance :
}
\author{
Damien Tichit pour Sia Partners
}
