% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiALM.R
\docType{package}
\name{SiALM}
\alias{SiALM}
\alias{SiALM-package}
\title{SiALM est un package permettant de calculer un best-estimate pour une compagnie d'assurance proposant les prosuits suivants : contrats d'epargne.}
\description{
SiALM est un package permettant de calculer un best-estimate pour une compagnie d'assurance proposant les prosuits suivants : contrats d'epargne.
}
\author{
Damien Tichit <damien.tichit@sia-partners.com>
}
