% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System-class.R
\docType{class}
\name{System}
\alias{System}
\title{Classe \code{System}}
\description{
Cette classe regroupe les actifs et les passifs d'une compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{passif}}{est un objet de type \code{\link{Passif}}.}

\item{\code{actif}}{est un objet de type \code{\link{Actif}}.}

\item{\code{taux_pb}}{est une \code{list} contenant les taux de pb contractuels.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
