% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabMorta-class.R
\docType{class}
\name{TabMorta}
\alias{TabMorta}
\title{Classe \code{TabMorta}}
\description{
Cette classe represente une table de mortalite.
}
\section{Slots}{

\describe{
\item{\code{table}}{est un objet de type \code{\link{data.frame}} contenant la table de mortalite.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
