% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabRachat-class.R
\docType{class}
\name{TabRachat}
\alias{TabRachat}
\title{Classe \code{TabRachat}}
\description{
Cette classe represente une table de rachat
}
\section{Slots}{

\describe{
\item{\code{table}}{est un objet de type \code{\link{data.frame}} contenant la table de rachat}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
