% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tresorerie-class.R
\docType{class}
\name{Tresorerie}
\alias{Tresorerie}
\title{Classe \code{Tresorerie}}
\description{
Cette classe represente la tresorerie de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{ptf}}{est un objet de type \code{\link{data.frame}} contenant les donnees relatives au portfeuille.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
