% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-update_esg.R
\docType{methods}
\name{update_esg}
\alias{update_esg}
\title{Fonction \code{update_esg}.}
\usage{
update_esg(alm, num_simu)
}
\arguments{
\item{alm}{est un objet de type \code{\link{ALM}}.}

\item{num_simu}{est un \code{integer} representant le numero de simulation sur lequel on travaille.}
}
\description{
Cette fonction permet de mettre a jour les donnees de l'ESG dans la classe \code{\link{HypActif}}.
Elle met les donnees a jour pour une simulation.
}
\author{
Damien Tichit pour Sia Partners
}
