% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tresorerie-update.R
\docType{methods}
\name{update_tresorerie}
\alias{update_tresorerie}
\title{Fonction \code{update_tresorerie}.}
\usage{
update_tresorerie(tresorerie, montant)
}
\arguments{
\item{tresorerie}{est un objet de type \code{\link{Tresorerie}}}

\item{montant}{est un \code{numeric}}
}
\description{
Cette fonction permet de mettre a jour la tresorerie : credit et debit.
}
\author{
Damien Tichit pour Sia Partners
}
