% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-vieillissement.R
\docType{methods}
\name{vieillissement_obligation}
\alias{vieillissement_obligation}
\title{Fonction \code{vieillissement_obligation}}
\usage{
vieillissement_obligation(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet de vieillir un portfeuille obligataire : mise a jour de la maturite residuelle et vente des obligations arrivees a maturite.
}
\author{
Damien Tichit pour Sia Partners
}
