% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-vieilissement.R
\docType{methods}
\name{vieillissement_ppe}
\alias{vieillissement_ppe}
\title{Fonction \code{vieillissement_ppe}.}
\usage{
vieillissement_ppe(ppe)
}
\arguments{
\item{ppe}{est un objet de type \code{\link{PPE}}.}
}
\description{
Cette fonction permet de vieillir d'une annee l'objet \code{\link{PPE}}
}
\author{
Damien Tichit pour Sia Partners
}
