% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-vieillissement.R
\docType{methods}
\name{vieillissement_ptf_actif}
\alias{vieillissement_ptf_actif}
\title{Fonction \code{vieillissement_ptf_actif}}
\usage{
vieillissement_ptf_actif(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}
}
\description{
Cette fonction permet de vieillir les differents composants d'un portfeuille d'actif.
}
\author{
Damien Tichit pour Sia Partners
}
