% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-calc_be.R
\docType{methods}
\name{calc_be}
\alias{calc_be}
\title{Fonction \code{calc_be}.}
\usage{
calc_be(alm, parallel = FALSE, nb_core = 1L)
}
\arguments{
\item{alm}{est un objet de type \code{\link{ALM}} contenant l'ensemble des donnees.}

\item{parallel}{est une valeur \code{logical}. Lorsque cet argument est a \code{TRUE}, les calculs sont parallelises.}

\item{nb_core}{est une valeur \code{integer} qui indique le nombre de coeurs utilises lorsque les calculs sont parallelises.}
}
\description{
Cette fonction est une fonction centrale du package. Elle permet en effet de calculer un BEL.
}
\details{
Il est possible paralleliser les calculs afin d'accelerer le calcul d'un best-estimate.
}
\author{
Damien Tichit pour Sia Partners
}
