% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPE-dotation.R
\docType{methods}
\name{dotation_ppe}
\alias{dotation_ppe}
\title{Fonction \code{dotation_ppe}.}
\usage{
dotation_ppe(ppe, montant)
}
\arguments{
\item{ppe}{est un objet de type \code{\link{PPE}}.}

\item{montant}{est un \code{numeric} representant le montant a doter.}
}
\description{
Cette fonction permet de doter la PPE. Le montant est dote sur la 1ere annee de reserve.
}
\author{
Damien Tichit pour Sia Partners
}
