% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE-dotation.R, R/ReserveCapi-dotation.R
\docType{methods}
\name{dotation_reserve_capi}
\alias{dotation_reserve_capi}
\alias{dotation_reserve_capi}
\title{Fonction \code{dotation_pre}.}
\usage{
dotation_pre(pre, pmvl)

dotation_reserve_capi(reserve_capi, pmvr)
}
\arguments{
\item{pmvl}{est une \code{list} contenant les differentes PMVL par classes d'actifs modelisees.}

\item{reserve_capi}{est un objet de type \code{\link{PRE}}.}

\item{pmvr}{est un \code{numeric} representant le montant les plus ou moins value resalisees a doter.}

\item{reserve_capi}{est un objet de type \code{\link{ReserveCapi}}.}
}
\description{
Cette fonction permet de doter la PRE.

Cette fonction permet de doter la Reserve de capitalisation. Sa modelisation reflete sa definition (Article A333.3).
}
\details{
En cas de plus value (montant > 0) : dotation totale.
En cas de moins value (montant < 0) : reprise dans la reserve de capitalisation, sous reserve qu'il soit encore positif.
}
\author{
Damien Tichit pour Sia Partners

Damien Tichit pour Sia Partners
}
