% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epargne-class.R
\docType{class}
\name{Epargne}
\alias{Epargne}
\title{Classe \code{Epargne}}
\description{
Cette classe represente le portefeuille des contrats epargne de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{ptf}}{est un objet de type \code{\link{data.frame}} contenant les donnees relatives au portfeuille.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
