% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-eval_frais_fin.R
\docType{methods}
\name{eval_frais_fin}
\alias{eval_frais_fin}
\title{Fonction \code{eval_frais_fin}}
\usage{
eval_frais_fin(ptf_actif, hyp_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}

\item{hyp_actif}{est un objet de type \code{\link{HypActif}}.}
}
\description{
Cette fonction permet d'evaluer les frais relatifs au portfeuille actif d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
