% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immobilier-gestion.R
\docType{methods}
\name{gestion_immobilier}
\alias{gestion_immobilier}
\title{Fonction \code{gestion_immobilier}}
\usage{
gestion_immobilier(immobilier, hyp_actif, an)
}
\arguments{
\item{immobilier}{est un objet de type \code{\link{Immobilier}}.}

\item{hyp_actif}{est un objet de type \code{\link{HypActif}}.}

\item{an}{est un \code{integer} reprensentant l'annee sur laquelle on travaille.}
}
\description{
Cette fonction permet de gerer un portfeuille immobilier : recolte des loyers, recalcul des VM
}
\author{
Damien Tichit pour Sia Partners
}
