% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypPassif-class.R
\docType{class}
\name{HypPassif}
\alias{HypPassif}
\title{Classe \code{HypPassif}}
\description{
Cette classe aggrege l'ensemble des hypotheses relatives au passif d'une compagnie d'assurance : tables de rachat, de mortalite...
}
\section{Slots}{

\describe{
\item{\code{tab_morta_h}}{est un objet de la classe \code{\link{TabMorta}} contenant la table de mortalite pour les hommes.}

\item{\code{tab_morta_f}}{est un objet de la classe \code{\link{TabMorta}} contenant la table de mortalite pour les hommes.}

\item{\code{tab_rachat_tot}}{est un objet de la classe \code{\link{TabRachat}} contenant la table modelisant les rachats totaux.}

\item{\code{tab_rachat_part}}{est un objet de la classe \code{\link{TabRachat}} contenant la table modelisant les rachats partiels.}

\item{\code{rachat_conj}}{est un objet de la classe \code{\link{RachatConj}} contenant les parametres modelisant les rachats conjoncturels.}

\item{\code{calc_proba}}{est un objet \code{logical} qui indique si les probabilites doivent etre calculees.}

\item{\code{prop_pb}}{est un \code{data.frame} qui indique la proportion a attribue dans le cas ou les besoins ont ete assouvis (apres besoins contractuels et cible).}

\item{\code{cible}}{est une \code{list} contenant les differents taux cibles par produits.}
}}

\seealso{
La mise a jour de l'attribut \code{cible} : \code{\link{update_esg}}
}
\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
