% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypPassif-class.R
\docType{class}
\name{HypPassif}
\alias{HypPassif}
\title{Classe \code{HypPassif}}
\description{
Cette classe aggrege l'ensemble des hypotheses relatives au passif d'une compagnie d'assurance : tables de rachat, de mortalite...
}
\section{Slots}{

\describe{
\item{\code{tab_morta_h}}{est un objet de la classe \code{\link{TabMorta}} contenant la table de mortalite pour les hommes.}

\item{\code{tab_morta_f}}{est un objet de la classe \code{\link{TabMorta}} contenant la table de mortalite pour les hommes.}

\item{\code{tab_rachat_tot}}{est un objet de la classe \code{\link{TabRachat}} contenant la table modelisant les rachats totaux.}

\item{\code{tab_rachat_part}}{est un objet de la classe \code{\link{TabRachat}} contenant la table modelisant les rachats partiels.}

\item{\code{rachat_conj}}{est un objet de la classe \code{\link{RachatConj}} contenant les parametres modelisant les rachats conjoncturels}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
