% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM-initialisation.R
\docType{methods}
\name{initialisation_alm}
\alias{initialisation_alm}
\title{Fonction \code{initialisation_alm}.}
\usage{
initialisation_alm(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaire
pour la construction de l'objet.}
}
\description{
Cette fonction permet d'initialiser un objet de type \code{\link{ALM}} : chargement des donnees, aggregation, calcul des probas...
}
\seealso{
Construction d'un objet de type \code{\link{ALM}} : \code{\link{load_alm}}.

Premiere aggregation d'un objet de type \code{\link{ALM}} : \code{\link{aggregation_alm}}.

Seconde aggregation pour les passifs \code{\link{ALM}} : \code{\link{aggregation_passif_2}}.
}
\author{
Damien Tichit pour Sia Partners
}
