% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actif-load.R
\docType{methods}
\name{load_actif}
\alias{load_actif}
\title{Fonction \code{load_actif}.}
\usage{
load_actif(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{Actif}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{Actif}} necessite la creation d'e deux autres objets'un autre objet : \code{\link{PTFActif}}.
}
\seealso{
Construction d'un objet de type \code{\link{PTFActif}} : \code{\link{load_ptf_actif}}.
}
\author{
Damien Tichit pour Sia Partners
}
