% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFCible-load.R
\docType{methods}
\name{load_ptf_cible}
\alias{load_ptf_cible}
\title{Fonction \code{load_ptf_cible}.}
\usage{
load_ptf_cible(address)
}
\arguments{
\item{address}{est un objet de type \code{character} indiquant le dossier dans lequel se situe l'ensemble des donnees necessaires
pour la construction de l'objet.}
}
\description{
Cette fonction permet de charger les donnees pour un objet de type \code{\link{PTFCible}}. Les donnees auront ete prealablement disposees dans
une architecture propre a \code{SiALM}.
}
\details{
La creation d'un objet \code{\link{PTFCible}} necessite la creation d'un autre objet : \code{\link{Obligation}}.
}
\seealso{
Construction d'un objet de type \code{\link{Obligation}} : \code{\link{load_obligation_cible}}.
}
\author{
Damien Tichit pour Sia Partners
}
