% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-class.R
\docType{class}
\name{Passif}
\alias{Passif}
\title{Classe \code{Passif}}
\description{
Cette classe aggrege l'ensemble des donnees relatives au passif de la compagnie d'assurance : hypotheses, portefeuille
}
\section{Slots}{

\describe{
\item{\code{ptf_passif}}{est un objet de la classe \code{\link{PTFPassif}} representant le portfeuille passif.}

\item{\code{hyp_passif}}{est un objet de la classe \code{\link{HypPassif}} contenant l'ensemble des hypotheses su passif.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
