% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actif-proj_1an.R
\docType{methods}
\name{proj_1an_actif}
\alias{proj_1an_actif}
\title{Fonction \code{proj_1an_actif}.}
\usage{
proj_1an_actif(actif)
}
\arguments{
\item{actif}{est un objet de type \code{\link{Actif}}.}
}
\description{
Cette fonction permet de projeter horizon 1 an l'actif d'une compagnie d'assurance.
}
\author{
Damien Tichit pour Sia Partners
}
