% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passif-proj_1an.R
\docType{methods}
\name{proj_1an_passif}
\alias{proj_1an_passif}
\title{Fonction \code{proj_1an_passif}.}
\usage{
proj_1an_passif(passif, an)
}
\arguments{
\item{passif}{est un objet de type \code{Passif}.}

\item{an}{est un \code{integer}.}
}
\description{
Cette fonction permet de projeter horizon 1 an un portefeuille passif.
}
\author{
Damien Tichit pour Sia Partners
}
