% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actif-rebalancement.R
\docType{methods}
\name{rebalancement_actif}
\alias{rebalancement_actif}
\title{Fonction \code{rebalancement_actif}}
\usage{
rebalancement_actif(actif)
}
\arguments{
\item{actif}{est un objet de type \code{Actif}.}
}
\description{
Cette fonction permet de rebalancer le portfeuille d'actif vers le portfeuile cible.
}
\author{
Damien Tichit pour Sia Partners
}
