% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obligation-revalo.R
\docType{methods}
\name{revalo_obligation}
\alias{revalo_obligation}
\title{Fonction \code{revalo_obligation}}
\usage{
revalo_obligation(obligation)
}
\arguments{
\item{obligation}{est un objet de type \code{\link{Obligation}}.}
}
\description{
Cette fonction permet de revaloriser le portefeuille obligataire.
}
\author{
Damien Tichit pour Sia Partners
}
