% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTFActif-revalo.R
\docType{methods}
\name{revalo_ptf_actif}
\alias{revalo_ptf_actif}
\title{Fonction \code{revalo_ptf_actif}}
\usage{
revalo_ptf_actif(ptf_actif)
}
\arguments{
\item{ptf_actif}{est un objet de type \code{\link{PTFActif}}.}
}
\description{
Cette fonction permet de projeter de revaloriser le portfeuille financier. Elle calcule egalement les plus ou moins value latentes (PMVL) engendrees.
}
\author{
Damien Tichit pour Sia Partners
}
