% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tresorerie-class.R
\docType{class}
\name{Tresorerie}
\alias{Tresorerie}
\title{Classe \code{Tresorerie}}
\description{
Cette classe represente la tresorerie de la compagnie d'assurance.
}
\section{Slots}{

\describe{
\item{\code{solde}}{est un objet \code{numeric} indiquant le solde.}
}}

\author{
Damien Tichit pour Sia Partners
}
\keyword{classes}
